sequencer_utils

Function contract_send

pub async fn contract_send<T, P, C>(
    call: &SolCallBuilder<T, P, C>,
) -> Result<(TransactionReceipt, u64), Error>
where
    P: Provider,
    C: SolCall,

send a transaction and wait for confirmation before returning the tx receipt and block included.

NOTE:

wait_for_transaction_to_be_mined is removed thanks to alloy’s better builtin PendingTransaction await

DON’T use this if you want parse the exact revert reason/type, since this func will only give err msg like: “custom error 0x23b0db14”, instead, follow https://docs.rs/alloy/0.12.5/alloy/contract/enum.Error.html#method.as_decoded_interface_error to pattern-match err type