sequencer_utils::ser
Module string_or_integer

Deserialize a type from either a string or integer in human-readable encodings.

This serialization module can be used with any FromStringOrInteger type. It is usually used only indirectly by the expansion of the [impl_string_or_integer] macro.

Enums

StringOrInteger 🔒

Functions

deserialize

serialize