Crate sequencer_utils

Modules

deployer

logging

ser

test_utils

Macros

impl_serde_from_string_or_integer

Deserialize a type from either a string or integer in human-readable encodings.

impl_to_fixed_bytes

Implement to_fixed_bytes for wrapped types

Structs

Anvil

Convenient interfaces for using anvil command which runs a local blockchain Similar to [AnvilInstance][https://docs.rs/ethers/latest/ethers/core/utils/struct.AnvilInstance.html], with more useful methods

AnvilOptions

Functions

commitment_to_u256

converting a keccak256-based structured commitment (32 bytes) into type U256

contract_send

send a transaction and wait for confirmation before returning the tx receipt and block included.

init_signer

Prepare a SignerMiddleware by connecting to a provider and initiating a local wallet. Returns a signer/client that can sign and send transactions to network.

u256_to_commitment

converting a U256 value into a keccak256-based structured commitment (32 bytes)

wait_for_http

wait_for_rpc

wait_for_transaction_to_be_mined 🔒

Type Aliases

NonceManager

Signer